% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coabundance.R
\name{as_coabundance}
\alias{as_coabundance}
\title{Convert coabundance objects}
\usage{
as_coabundance(x, ...)
}
\arguments{
\item{x}{object to transform to class coabundance. Supported classes: \code{pulsar.refit}, \code{tbl_df}, \code{rcorr}, \code{spiec_easi_sparcc_res}}
}
\description{
Convert coabundance objects
}
