% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{correlate}
\alias{correlate}
\title{Coabundance Analysis}
\usage{
correlate(data, method = "sparcc", ...)
}
\arguments{
\item{data}{matrix or data frame with abundance count data}

\item{method}{character of coabundance method. One of 'sparcc', 'mb', 'pearson', or 'spearman'}

\item{...}{arguments passed to selected function (one of \link[coabundance]{correlate_sparcc},
\link[coabundance]{correlate_mb}, \link[coabundance]{correlate_spearman}, or \link[coabundance]{correlate_pearson})}
}
\description{
Coabundance Analysis
}
\details{
The option \code{method} (\code{sparcc} by default) defines the method use to calculate the interaction between pairs of coabundant taxa:
\itemize{
\item \code{pearson}: Pearson correlation coefficient
\item \code{spearman}: Spearman's rank correlation coefficient
\item \code{mb}: Inverse Covariance based on \insertCite{mb}{coabundance} as implemented in \insertCite{spiec_easi}{coabundance}
\item \code{sparcc}: SparCC correlation based on \insertCite{sparcc}{coabundance} as implemented in \insertCite{spiec_easi}{coabundance}
or \insertCite{fastspar}{coabundance}
}
}
\references{
\insertRef{mb}{coabundance}

\insertRef{spiec_easi}{coabundance}

\insertRef{sparcc}{coabundance}

\insertRef{fastspar}{coabundance}
}
