% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{correlate_fastspar}
\alias{correlate_fastspar}
\title{Do fastspar coabundance}
\usage{
correlate_fastspar(
  data,
  iterations = 50,
  exclude_iterations = 10,
  bootstraps = 200,
  threads = getOption("mc.cores")
)
}
\arguments{
\item{data}{integer matrix of abundance count data. One sample per row and one taxon per column}
}
\description{
Do fastspar coabundance
}
