% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{correlate_sparcc}
\alias{correlate_sparcc}
\title{Coabundance analysis using SparCC}
\usage{
correlate_sparcc(implementation = "spiec_easi", ...)
}
\arguments{
\item{implementation}{Character indicating the implementation of SparCC algorithm to use. One of "fastspar", "spiec_easi"}

\item{...}{further arguments passed to the sparcc correlation functions}
}
\description{
Coabundance analysis using SparCC
}
