% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coabundance.R
\name{filter.coabundance}
\alias{filter.coabundance}
\title{Filter a coabundance object}
\usage{
filter.coabundance(
  x,
  max_pval = 0.05,
  min_abs_estimate = NULL,
  remove_isolated_nodes = TRUE,
  recalculate_topology = TRUE
)
}
\arguments{
\item{x}{coabundance object}

\item{max_pval}{maximum p value to keep an edge. Will be ignores if the p.value is not available}

\item{min_abs_estimate}{minimal absolute value of the estimate to keep an edge. This is to filter edges with a low effect size.}

\item{remove_isolated_nodes}{TRUE if nodes not being part of any edge after filtering should be removed, FALSE otherwise.}

\item{recalculate_topology}{TRUE if node topology metrics e.g. centrality scores should be recalculated after filtering, FALSE otherwise.}
}
\description{
Filter a coabundance object
}
