% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{correlate_fastspar}
\alias{correlate_fastspar}
\title{Coabundance Analysis Using SparCC as Implemented in Fastspar}
\usage{
correlate_fastspar(
  data,
  iterations = 50,
  exclude_iterations = 10,
  bootstraps = 200,
  threads = getOption("mc.cores")
)
}
\arguments{
\item{data}{integer matrix of abundance count data. One sample per row and one taxon per column}
}
\description{
This implementation is way faster than \code{correlate_spiec_easi_sparcc} but requires linux and the external shell command \code{fastspar}.
}
