% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{correlate_spiec_easi_sparcc}
\alias{correlate_spiec_easi_sparcc}
\title{Coabundance analysis using SparCC as implemented in SpiecEasi}
\usage{
correlate_spiec_easi_sparcc(
  data,
  iterations = 10,
  bootstraps = 200,
  threads = 1
)
}
\arguments{
\item{data}{integer matrix of abundance count data. One sample per row and one taxon per column}
}
\description{
Coabundance analysis using SparCC as implemented in SpiecEasi
}
